backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment Variables
python-dotenv>=1.0,<2.0

# CORS Headers
django-cors-headers>=4.3,<5.0

# Production Server
gunicorn>=21.2,<22.0

# Additional utilities
dj-database-url>=2.1,<3.0
